Restructured the cycle calculations and split the Python modules
